! numerical problems
The distribution would allow for theoretical values equal to zero, which
is not possible for the chi^2 test used here (without Yates correction).

The code now a) detects that case and skips the test (printing a
warning) and b) only generates non-null probabilities. This should
pretty much eliminate the problem altogether